+++ changelog.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - Add the ability to have minimum delay between short messages (e.g. callmebot signal)
 - Minor update for Deepsea controller switch state
 - More updates for ComAp controller
+- Update to gentankutil to add option to check battery and missed readings. See Add On page to enable this functionality.
 
 ## V1.19.00 -2023-11-09
 - Updates to genmonmain.sh, startgenmon.sh and genloader.sh for Debian bookworm.